# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Battle Rap API
  version: 1.0.0
  description: >
    REST API for managing online rap battles, tournaments, submissions, judging,
    and administration workflows.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/tournaments:
    get:
      tags: [Public]
      summary: List tournaments
      operationId: listTournaments
      parameters:
        - in: query
          name: status
          schema:
            type: string
          description: Filter by tournament status.
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: Page number (1-indexed).
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Page size.
      responses:
        '200':
          description: Paginated tournament list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentListResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/tournaments/{id}:
    get:
      tags: [Public]
      summary: Get tournament with rounds
      operationId: getTournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tournament details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament:
                    $ref: '#/components/schemas/Tournament'
                  rounds:
                    type: array
                    items:
                      $ref: '#/components/schemas/Round'
                required:
                  - tournament
                  - rounds
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/rounds/{id}/overview:
    get:
      tags: [Public]
      summary: Round overview with scoring summaries
      operationId: getRoundOverview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Round overview data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoundOverview'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/rounds/{id}:
    get:
      tags: [Public]
      summary: Get round with matches
      operationId: getRound
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Round with associated matches.
          content:
            application/json:
              schema:
                type: object
                properties:
                  round:
                    $ref: '#/components/schemas/Round'
                  matches:
                    type: array
                    items:
                      $ref: '#/components/schemas/MatchSummary'
                required:
                  - round
                  - matches
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/rounds/{id}/battles:
    get:
      tags: [Public]
      summary: List matches in a round with conditional winners
      operationId: listRoundBattles
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Round and matches.
          content:
            application/json:
              schema:
                type: object
                properties:
                  round:
                    $ref: '#/components/schemas/Round'
                  matches:
                    type: array
                    items:
                      $ref: '#/components/schemas/MatchWithMaskedWinner'
                required:
                  - round
                  - matches
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/battles:
    get:
      tags: [Public]
      summary: List public battles with participants
      operationId: listPublicBattles
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [current, finished]
          description: Filter to active or finished battles.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
          description: Maximum number of battles to return (default 20).
      responses:
        '200':
          description: Public battle listing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicBattleListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/battles/{id}:
    get:
      tags: [Public]
      summary: Get battle summary
      operationId: getBattle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Battle details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Battle'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/admin/roles/{userId}:
    post:
      tags: [Admin]
      summary: Set user role
      operationId: setUserRole
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleChangeRequest'
      responses:
        '200':
          description: Updated user roles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRolesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/admin/finalize/battles/{id}:
    post:
      tags: [Admin]
      summary: Finalize battle and refresh views
      operationId: finalizeBattle
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Battle finalized.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/admin/users:
    get:
      tags: [Admin]
      summary: List users with roles
      operationId: listUsersForAdmin
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: search
          schema:
            type: string
            description: Case-insensitive match against display name, email, or full name.
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, moderator, artist, judge, listener]
        - in: query
          name: sort
          schema:
            type: string
            enum: [created_at, -created_at, display_name, -display_name]
      responses:
        '200':
          description: Paginated user list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/admin/users/{id}:
    get:
      tags: [Admin]
      summary: Get user profile as admin
      operationId: getUserProfileAdmin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Full user profile with private fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/admin/audit-log:
    get:
      tags: [Admin]
      summary: List audit log entries
      operationId: listAuditLog
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: actor_id
          schema:
            type: string
            format: uuid
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: target_table
          schema:
            type: string
        - in: query
          name: target_id
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Audit log entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token issued by the API. Required for authenticated endpoints.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        Refresh token cookie issued during login/registration.
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentication required or token invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        status:
          type: integer
        field_errors:
          type: object
          additionalProperties:
            type: string
      example:
        type: https://httpstatuses.com/422
        title: Payload validation failed.
        detail: Payload validation failed.
        status: 422
        field_errors:
          email: Invalid email.
    Tournament:
      type: object
      required:
        - id
        - title
        - status
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          description: Tournament lifecycle status.
        registration_open_at:
          type: string
          format: date-time
          nullable: true
        submission_deadline_at:
          type: string
          format: date-time
          nullable: true
        judging_deadline_at:
          type: string
          format: date-time
          nullable: true
        public_at:
          type: string
          format: date-time
          nullable: true
    TournamentListResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Tournament'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    Round:
      type: object
      required:
        - id
        - tournament_id
        - kind
        - number
        - scoring
        - status
      properties:
        id:
          type: string
          format: uuid
        tournament_id:
          type: string
          format: uuid
        kind:
          type: string
          description: Round type (e.g. qualification, duel).
        number:
          type: integer
        scoring:
          type: string
          enum: [pass_fail, points, rubric]
        status:
          type: string
          description: Current round state.
        starts_at:
          type: string
          format: date-time
          nullable: true
        submission_deadline_at:
          type: string
          format: date-time
          nullable: true
        judging_deadline_at:
          type: string
          format: date-time
          nullable: true
        strategy:
          type: string
          nullable: true
    MatchSummary:
      type: object
      required:
        - id
        - round_id
        - status
      properties:
        id:
          type: string
          format: uuid
        round_id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        ends_at:
          type: string
          format: date-time
          nullable: true
        winner_match_track_id:
          type: string
          format: uuid
          nullable: true
    MatchWithMaskedWinner:
      allOf:
        - $ref: '#/components/schemas/MatchSummary'
...
    PublicBattleListResponse:
      type: object
      required:
        - battles
      properties:
        battles:
          type: array
          items:
            $ref: '#/components/schemas/PublicBattle'
    Battle:
      type: object
      required:
        - id
        - round_id
        - status
      properties:
        id:
          type: string
          format: uuid
        round_id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        ends_at:
          type: string
          format: date-time
          nullable: true
        winner_match_track_id:
          type: string
          format: uuid
          nullable: true
        engagement:
          $ref: '#/components/schemas/BattleEngagementSummary'
...
    BattleEngagementSummary:
      type: object
      required:
        - comments
      properties:
        comments:
          type: integer
          minimum: 0
...
    ProfileAvatar:
      type: object
      required:
        - key
        - url
      properties:
        key:
          type: string
          description: Internal storage key for the avatar asset.
        url:
          type: string
          format: uri
          description: CDN URL for the avatar asset.
...
    ProfileView:
      type: object
      required:
        - id
        - display_name
        - roles
        - created_at
        - updated_at
        - viewer_context
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        roles:
          type: array
          items:
            type: string
        avatar:
          $ref: '#/components/schemas/ProfileAvatar'
          nullable: true
        bio:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        viewer_context:
          $ref: '#/components/schemas/ProfileViewerContext'
        email:
          type: string
          format: email
          description: Present when the caller can view private information.
        age:
          type: integer
          minimum: 12
          maximum: 120
          nullable: true
        vk_id:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        socials:
          type: object
          additionalProperties: true
          description: Social handles map, available to privileged viewers.
...
    AdminUserListResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...
    AuditLogResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...
    RoleChangeRequest:
      type: object
      required:
        - op
        - role
      properties:
        op:
          type: string
          enum: [grant, revoke]
        role:
          type: string
          enum: [artist, judge, listener, moderator, admin]
    UserRolesResponse:
      type: object
      required:
        - user_id
        - roles
      properties:
        user_id:
          type: string
          format: uuid
        roles:
          type: array
          items:
            type: string
...
    JudgingRubricCriterion:
      type: object
      required:
        - key
        - name
      properties:
        key:
          type: string
        name:
          type: string
        weight:
          type: number
          nullable: true
        min_value:
          type: number
          nullable: true
        max_value:
          type: number
          nullable: true
        position:
          type: integer
          nullable: true
...
    RoundOverview:
      type: object
      required:
        - round
        - mode
        - summary
      properties:
        round:
          $ref: '#/components/schemas/Round'
        mode:
          type: string
          enum: [pass_fail, points, rubric]
        summary:
          type: object
          properties:
            total_submissions:
              type: integer
              nullable: true
            total_matches:
              type: integer
              nullable: true
            total_tracks:
              type: integer
              nullable: true
            total_reviews:
              type: integer
              nullable: true
            mode:
              type: string
        submissions:
          type: array
          nullable: true
          items:
            type: object
            required:
              - submission_id
              - participant_id
              - user_id
              - display_name
            properties:
              submission_id:
                type: string
                format: uuid
              participant_id:
                type: string
                format: uuid
              user_id:
                type: string
                format: uuid
              display_name:
                type: string
              status:
                type: string
              submitted_at:
                type: string
                format: date-time
                nullable: true
              lyrics:
                type: string
                nullable: true
              pass_count:
                type: integer
              fail_count:
                type: integer
              judge_count:
                type: integer
              total_score:
                type: number
              avg_score:
                type: number
                nullable: true
              total_reviews:
                type: integer
              audio:
                type: object
                properties:
                  key:
                    type: string
                  url:
                    type: string
                    format: uri
                  mime:
                    type: string
                  duration_sec:
                    type: number
                    nullable: true
              avatar:
                type: object
                nullable: true
                properties:
                  key:
                    type: string
                  url:
                    type: string
                    format: uri
        matches:
          type: array
          nullable: true
          items:
            type: object
            required:
              - id
              - status
              - participants
            properties:
              id:
                type: string
                format: uuid
              status:
                type: string
              starts_at:
                type: string
                format: date-time
                nullable: true
              ends_at:
                type: string
                format: date-time
                nullable: true
              winner_match_track_id:
                type: string
                format: uuid
                nullable: true
              participants:
                type: array
                items:
                  type: object
                  required:
                    - participant_id
                    - user_id
                    - display_name
                  properties:
                    participant_id:
                      type: string
                      format: uuid
                    user_id:
                      type: string
                      format: uuid
                    display_name:
                      type: string
                    seed:
                      type: integer
                      nullable: true
                    avg_total_score:
                      type: number
                      nullable: true
                    track:
                      type: object
                      nullable: true
                      properties:
                        id:
                          type: string
                          format: uuid
                        audio_key:
                          type: string
                          nullable: true
                        audio_url:
                          type: string
                          format: uri
                          nullable: true
                        mime:
                          type: string
                          nullable: true
                        duration_sec:
                          type: number
                          nullable: true
                        submitted_at:
                          type: string
                          format: date-time
                          nullable: true
                        lyrics:
                          type: string
                          nullable: true
        rubric:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/JudgingRubricCriterion'
...